CS210--Introduction-to-Data-Science--course-part-2
The objective of the third phase of the project is applying the machine learning methods that are learned in the course by using the statistical analysis which is done in Phase 2. This long journey involves three primary stages: Building two machine learning model, verifying these models and deciding which machine learning performs better and the reason why this model performs better. At the end of these three stages, the goal is building most appropriate machine learning model for statistical analysis.